src/test/java/com/abosen/test/v4/ClassPathBeanDefinitionScannerTest.java
package com.abosen.test.v4;

import com.abosen.beans.BeanDefinition;
import com.abosen.beans.factory.support.DefaultBeanFactory;
import com.abosen.context.annotation.ClassPathBeanDefinitionScanner;
import com.abosen.context.annotation.ScannedGenericBeanDefinition;
import com.abosen.core.annotation.AnnotationAttributes;
import com.abosen.core.type.AnnotationMetadata;
import com.abosen.stereotype.Component;
import org.junit.Assert;
import org.junit.Test;

/**
 * @author qiubaisen
 * @date 2018/7/24
 */
public class ClassPathBeanDefinitionScannerTest {

    @Test
    public void testParseScannedBean(){

        DefaultBeanFactory factory = new DefaultBeanFactory();

        String basePackages = "com.abosen.service.v4,com.abosen.dao.v4";

        ClassPathBeanDefinitionScanner scanner = new ClassPathBeanDefinitionScanner(factory);
        scanner.doScan(basePackages);


        String annotation = Component.class.getName();

        {
            BeanDefinition bd = factory.getBeanDefinition("petStore");
            Assert.assertTrue(bd instanceof ScannedGenericBeanDefinition);
            ScannedGenericBeanDefinition sbd = (ScannedGenericBeanDefinition)bd;
            AnnotationMetadata amd = sbd.getMetadata();


            Assert.assertTrue(amd.hasAnnotation(annotation));
            AnnotationAttributes attributes = amd.getAnnotationAttributes(annotation);
            Assert.assertEquals("petStore", attributes.get("value"));
        }
        {
            BeanDefinition bd = factory.getBeanDefinition("accountDao");
            Assert.assertTrue(bd instanceof ScannedGenericBeanDefinition);
            ScannedGenericBeanDefinition sbd = (ScannedGenericBeanDefinition)bd;
            AnnotationMetadata amd = sbd.getMetadata();
            Assert.assertTrue(amd.hasAnnotation(annotation));
        }
        {
            BeanDefinition bd = factory.getBeanDefinition("itemDao");
            Assert.assertTrue(bd instanceof ScannedGenericBeanDefinition);
            ScannedGenericBeanDefinition sbd = (ScannedGenericBeanDefinition)bd;
            AnnotationMetadata amd = sbd.getMetadata();
            Assert.assertTrue(amd.hasAnnotation(annotation));
        }
    }
}
